requirements.txt
fastapi
uvicorn
langchain
langchain-openai
langchain-community
langchain-text-splitters
pypdf
faiss-cpu
python-dotenv